Dp part2 fix1 after review

+++ main-service/src/main/java/ru/practicum/service/category/service/admin/CategoryServiceAdminDefault.java
@@ -1,14 +1,13 @@
 package ru.practicum.service.category.service.admin;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.service.category.dto.CategoryDto;
 import ru.practicum.service.category.dto.NewCategoryDto;
 import ru.practicum.service.category.model.Category;
 import ru.practicum.service.category.repository.CategoryRepository;
+import ru.practicum.service.event.repository.EventRepository;
 import ru.practicum.service.exception.CategoryHasRelatedEventsException;
 import ru.practicum.service.exception.CategoryIsNotInRepositoryException;
 import ru.practicum.service.exception.CategoryNameNotUniqueException;
@@ -16,38 +15,42 @@ import ru.practicum.service.exception.CategoryNameNotUniqueException;
 import static ru.practicum.service.category.dto.mapper.CategoryMapper.*;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class CategoryServiceAdminDefault implements CategoryServiceAdmin {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto create(NewCategoryDto newCategoryDto) {
-        try {
-            return toCategoryDto(categoryRepository.save(toCategory(newCategoryDto)));
-        } catch (DataIntegrityViolationException e) {
-            throw new CategoryNameNotUniqueException("Trying to add a category that already exists in the database");
-        }
+
+        String name = newCategoryDto.getName();
+
+        checkIfCategoryNameIdNotUnique(name);
+
+        Category category = toCategory(newCategoryDto);
+
+        return toCategoryDto(categoryRepository.save(category));
     }
 
     @Override
     public void deleteById(Long catId) {
-        try {
-            categoryRepository.deleteById(catId);
-        } catch (EmptyResultDataAccessException e) {
-            throw new CategoryIsNotInRepositoryException("Category to delete not found in repository");
-        } catch (DataIntegrityViolationException e) {
-            throw new CategoryHasRelatedEventsException("The category cannot be deleted: " +
+
+        findCategoryByIdIfExists(catId);
+
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new CategoryHasRelatedEventsException("The category (id: " + catId + ") cannot be deleted: " +
                     "it has events associated with it");
         }
+
+        categoryRepository.deleteById(catId);
     }
 
     @Override
-    @Transactional
     public CategoryDto updateById(Long catId, NewCategoryDto newCategoryDtoForPatch) {
 
-        Category categoryOldVersion = categoryRepository.findById(catId)
-                .orElseThrow(() -> new CategoryIsNotInRepositoryException("Category for update not found in repository"));
+        Category categoryOldVersion = findCategoryByIdIfExists(catId);
 
         String newName = newCategoryDtoForPatch.getName();
 
@@ -55,12 +58,22 @@ public class CategoryServiceAdminDefault implements CategoryServiceAdmin {
             return toCategoryDto(categoryOldVersion);
         }
 
-        if (categoryRepository.existsByName(newName)) {
-            throw new CategoryNameNotUniqueException("Category name must be unique");
-        }
+        checkIfCategoryNameIdNotUnique(newName);
 
         categoryOldVersion.setName(newName);
 
         return toCategoryDto(categoryRepository.save(categoryOldVersion));
+    }
+
+    private Category findCategoryByIdIfExists(Long catId) {
+        return categoryRepository.findById(catId)
+                .orElseThrow(() -> new CategoryIsNotInRepositoryException("Category with id: " + catId +
+                        " is not in repository"));
+    }
+
+    private void checkIfCategoryNameIdNotUnique(String name) {
+        if (categoryRepository.existsByName(name)) {
+            throw new CategoryNameNotUniqueException("Category name: " + name + " is not unique");
+        }
     }
 }
